% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseLM.R
\name{sparselm}
\alias{sparselm}
\title{Nonlinear Least Squares Fit with Sparse Levenberg-Marquardt Algorithm}
\usage{
sparselm(
  p,
  x,
  func,
  fjac,
  Jnnz,
  JtJnnz = -1,
  nconvars = 0,
  itmax = 100,
  opts = sparselm.opts(),
  dif = FALSE,
  ...
)
}
\arguments{
\item{p}{initial parameter estimates length nvars}

\item{x}{measurement vector length nobs}

\item{func}{functional relation describing measurements given a parameter vector p, returning vector length nobs}

\item{fjac}{function to supply the nonzero pattern of the sparse Jacobian of func and optionally evaluate it at p, returning nobs by nvars dgCMatrix}

\item{Jnnz}{number of nonzeros for the Jacobian J}

\item{JtJnnz}{number of nonzeros for the product J^t*J, -1 if unknown}

\item{itmax}{maximum number of iterations}

\item{opts}{minim. options \code{mu, epsilon1, epsilon2, epsilon3, delta, spsolver}, not currently implemented!}

\item{dif}{logical indicating whether to use finite differences}

\item{...}{additional arguments passed to func and fjac}
}
\value{
list with four elements: par, niter, info, and term
}
\description{
Nonlinear Least Squares Fit with Sparse Levenberg-Marquardt Algorithm
}
